Refactored folders and files
