+++ src/components/Header.js
@@ -1,4 +1,4 @@
-import '../Styles/header.css';
+import '../styles/header.css';
 function Header(){
     return (
         <>
