+++ src/components/navbar.js
@@ -1,5 +1,5 @@
-import React, { useState } from 'react';
-import './navbar.css'
+import React from 'react';
+import '../styles/navbar.css'
 const Navbar = () => {
     return (
     <div id='mainNavbarDiv'>
