+++ src/styles/container.css
@@ -1,6 +1,6 @@
 .container1{
     width: 20%;
-    height: 145vh;
+    height: 100vh;
     background-color: rgb(28, 36, 52);
     display: block;
 }
