+++ src/styles/dashboard.css
@@ -6,7 +6,7 @@
     flex: 1;
 }
 
-.container {
+#chartContainer {
 
     margin-top: 5%;
